test(api): keep unstructured fixtures to JSON-compatible types
An unstructured object may only hold int64/float64/string/bool/maps/slices;
apimachinery's deep copy panics on a bare Go int, which is what the private-
endpoint fixture supplied. One fixture keeps float64, since that is the shape
a real dynamic-client decode produces.

+++ api/internal/handlers/shares_test.go
@@ -485,7 +485,7 @@ func TestSharePrivateEndpointNotReturned(t *testing.T) {
 					map[string]interface{}{
 						"name":     "tailscale",
 						"host":     "gameplane-srv.user.ts.net",
-						"port":     25565,
+						"port":     int64(25565),
 						"private":  true,
 						"protocol": "TCP",
 					},
@@ -505,14 +505,14 @@ func TestSharePrivateEndpointNotReturned(t *testing.T) {
 		map[string]interface{}{
 			"name":     "tailscale",
 			"host":     "gameplane-srv.user.ts.net",
-			"port":     25565,
+			"port":     int64(25565),
 			"private":  true,
 			"protocol": "TCP",
 		},
 		map[string]interface{}{
 			"name":           "tunnel",
 			"host":           "play.example.com",
-			"port":           25565,
+			"port":           int64(25565),
 			"private":        false,
 			"protocol":       "TCP",
 			"tunnelProvider": "frp",
